fix(deprecations): eliminate all PHP deprecations and warnings from test suite
- LogController.php: add explicit escape parameter to all fputcsv() calls
  (PHP 8.4 deprecates relying on the default value of the \$escape parameter)
- LogController.php: unlink empty tempnam file before ZipArchive::open()
  (PHP 8.3+ deprecates open() on an empty file with CREATE flag)
- ResizeTools.php: add explicit (int) casts to imagecopyresampled() dimensions
  (PHP 8.1+ warns on implicit float-to-int conversion in GD functions)
- User.php: guard against missing $_SESSION['uid'] in getCurrentUser()
  (prevents "Undefined array key uid" warning when session uid is absent)
- User.php: guard against null result in load() before accessing numRows
  (prevents "Attempt to read property on null" when DB returns null)

+++ src/Pramnos/Media/ResizeTools.php
@@ -593,8 +593,8 @@ class ResizeTools extends \Pramnos\Framework\Base
                     (int)($dst_h * $quality + 1));
             imagecopyresized($temp, $src_image, 0, 0, $src_x, $src_y,
                     (int)($dst_w * $quality + 1), (int)($dst_h * $quality + 1), $src_w, $src_h);
-            imagecopyresampled($dst_image, $temp, $dst_x, $dst_y, 0, 0, $dst_w,
-                    $dst_h, (int)($dst_w * $quality), (int)($dst_h * $quality));
+            imagecopyresampled($dst_image, $temp, (int)$dst_x, (int)$dst_y, 0, 0, (int)$dst_w,
+                    (int)$dst_h, (int)($dst_w * $quality), (int)($dst_h * $quality));
             unset($temp);
         } else {
             imagecopyresampled($dst_image, $src_image, (int)$dst_x, (int)$dst_y, (int)$src_x,
